+++ x-pack/platform/test/cases_api_integration/spaces_only/config.ts
@@ -12,4 +12,8 @@ export default createTestConfig('spaces_only', {
   license: 'trial',
   ssl: false,
   testFiles: [require.resolve('./tests/trial')],
+  // The trial suite includes templates / field-definitions coverage that
+  // requires the templates feature. Pin it ON explicitly so the suite is
+  // deterministic regardless of the plugin default (mirrors `config_trial.ts`).
+  kbnServerArgs: ['--xpack.cases.templates.enabled=true'],
 });
