+++ x-pack/platform/test/functional/services/cases/navigation.ts
@@ -44,6 +44,26 @@ export function CasesNavigationProvider({ getPageObject, getService }: FtrProvid
       });
     },
 
+    /**
+     * Navigates to the v2 templates list page (`configure/templates`).
+     * Only reachable when `xpack.cases.templates.enabled` is ON.
+     */
+    async navigateToTemplatesPage(app: string = 'cases') {
+      // The cases app pathname already ends with a trailing slash, so the
+      // sub-path is passed without a leading slash to avoid a double slash.
+      await common.navigateToUrlWithBrowserHistory(app, 'configure/templates');
+      await testSubjects.existOrFail('cases-app');
+    },
+
+    /**
+     * Navigates to the v2 field library page (`configure/field_library`).
+     * Only reachable when `xpack.cases.templates.enabled` is ON.
+     */
+    async navigateToFieldLibraryPage(app: string = 'cases') {
+      await common.navigateToUrlWithBrowserHistory(app, 'configure/field_library');
+      await testSubjects.existOrFail('cases-app');
+    },
+
     async navigateToSingleCase(app: string = 'cases', caseId: string, tabId?: string) {
       const search = tabId != null ? `?tabId=${tabId}` : '';
       await common.navigateToUrlWithBrowserHistory(app, caseId, search);
