+++ x-pack/platform/test/functional_with_es_ssl/apps/cases/group1/create_case_form.ts
@@ -159,6 +159,10 @@ export default ({ getService, getPageObject }: FtrProviderContext) => {
           return this.skip();
         }
 
+        // Templates v2 hides the legacy inline custom fields on Create behind a switch;
+        // reveal it so this legacy flow is exercised whether or not templates is on.
+        await cases.common.showLegacyCustomFields('cases');
+
         const customFields = [
           {
             key: 'valid_key_1',
@@ -219,6 +223,10 @@ export default ({ getService, getPageObject }: FtrProviderContext) => {
         if (await cases.common.isRedesignEnabled()) {
           return this.skip();
         }
+
+        // Templates v2 hides the legacy inline custom fields on Create behind a switch;
+        // reveal it so this legacy flow is exercised whether or not templates is on.
+        await cases.common.showLegacyCustomFields('cases');
 
         const customFields = [
           {
