+++ x-pack/platform/test/plugin_api_integration/config.ts
@@ -45,6 +45,10 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
         '--xpack.cases.analyticsV2.enabled=true',
         '--xpack.task_manager.monitored_aggregated_stats_refresh_rate=5000',
         '--xpack.task_manager.invalidate_api_key_task.removalDelay="1s"',
+        // Pin the Cases templates flag ON so the `cases_templates_v2_migration`
+        // task type is registered, matching the `check_registered_task_types` guard
+        // list regardless of the plugin default.
+        '--xpack.cases.templates.enabled=true',
         `--xpack.stack_connectors.enableExperimental=${JSON.stringify([
           'crowdstrikeConnectorOn',
           'microsoftDefenderEndpointOn',
