+++ x-pack/solutions/observability/test/serverless/functional/test_suites/cases/configure.ts
@@ -15,29 +15,32 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
   const svlCommonPage = getPageObject('svlCommonPage');
   const svlObltNavigation = getService('svlObltNavigation');
   const testSubjects = getService('testSubjects');
+  const browser = getService('browser');
   const cases = getService('cases');
   const svlCases = getService('svlCases');
   const toasts = getService('toasts');
   const retry = getService('retry');
-  const find = getService('find');
-  const comboBox = getService('comboBox');
+
+  const navigateToConfigure = async () => {
+    await svlObltNavigation.navigateToLandingPage();
+    await svlCommonNavigation.sidenav.clickLink({ deepLinkId: 'observability-overview:cases' });
+    await header.waitUntilLoadingHasFinished();
+
+    await cases.navigation.clickHeaderMenuItem('configure-case-button');
+    await header.waitUntilLoadingHasFinished();
+
+    await retry.waitFor('the configuration page to load', async () => {
+      return (
+        (await testSubjects.exists('case-configure-title')) ||
+        (await testSubjects.exists('cases-redesign-settings-panel'))
+      );
+    });
+  };
 
   describe('Configure Case', function () {
     before(async () => {
       await svlCommonPage.loginWithPrivilegedRole();
-      await svlObltNavigation.navigateToLandingPage();
-      await svlCommonNavigation.sidenav.clickLink({ deepLinkId: 'observability-overview:cases' });
-      await header.waitUntilLoadingHasFinished();
-
-      await cases.navigation.clickHeaderMenuItem('configure-case-button');
-      await header.waitUntilLoadingHasFinished();
-
-      await retry.waitFor('the configuration page to load', async () => {
-        return (
-          (await testSubjects.exists('case-configure-title')) ||
-          (await testSubjects.exists('cases-redesign-settings-panel'))
-        );
-      });
+      await navigateToConfigure();
     });
 
     after(async () => {
@@ -71,143 +74,41 @@ export default ({ getPageObject, getService }: FtrProviderContext) => {
       });
     });
 
-    describe('Custom fields', function () {
-      before(async function () {
-        // The redesign settings page does not manage custom fields (moved to the templates v2 UI).
-        if (await cases.common.isRedesignEnabled()) {
-          this.skip();
-        }
+    // With the templates feature flag pinned ON for this suite, custom fields and
+    // templates are managed on the dedicated v2 templates / field-library pages,
+    // not inline on the Case Settings page. The legacy in-page flow is covered by
+    // the flag-OFF stateful `configure_legacy.ts` suite.
+    describe('Templates page (v2)', function () {
+      before(async () => {
+        await navigateToConfigure();
+        // The cases app pathname already ends with `configure`, so append the
+        // sub-path after stripping any query string / hash.
+        const configureUrl = (await browser.getCurrentUrl()).split(/[?#]/)[0].replace(/\/$/, '');
+        await browser.get(`${configureUrl}/templates`);
+        await header.waitUntilLoadingHasFinished();
       });
 
-      it('adds a custom field', async () => {
-        await testSubjects.existOrFail('custom-fields-form-group');
-        await common.clickAndValidate('add-custom-field', 'common-flyout');
-
-        await testSubjects.setValue('custom-field-label-input', 'Summary');
-
-        await testSubjects.setCheckbox('text-custom-field-required-wrapper', 'check');
-
-        await testSubjects.click('common-flyout-save');
-        expect(await testSubjects.exists('euiFlyoutCloseButton')).to.be(false);
-
-        await testSubjects.existOrFail('custom-fields-list');
-
-        expect(await testSubjects.getVisibleText('custom-fields-list')).to.be('Summary\nText');
+      it('renders the templates list page with the empty prompt', async () => {
+        await testSubjects.existOrFail('templates-table-empty-prompt-no-templates');
+        await testSubjects.existOrFail('templates-table-add-template');
       });
 
-      it('edits a custom field', async () => {
-        await testSubjects.existOrFail('custom-fields-form-group');
-        const textField = await find.byCssSelector('[data-test-subj*="-custom-field-edit"]');
-
-        await textField.click();
-
-        const input = await testSubjects.find('custom-field-label-input');
-
-        await input.type('!!!');
-
-        await testSubjects.click('common-flyout-save');
-        expect(await testSubjects.exists('euiFlyoutCloseButton')).to.be(false);
-
-        await testSubjects.existOrFail('custom-fields-list');
-
-        expect(await testSubjects.getVisibleText('custom-fields-list')).to.be('Summary!!!\nText');
-      });
-
-      it('deletes a custom field', async () => {
-        await testSubjects.existOrFail('custom-fields-form-group');
-        const deleteButton = await find.byCssSelector('[data-test-subj*="-custom-field-delete"]');
-
-        await deleteButton.click();
-
-        await testSubjects.existOrFail('confirm-delete-modal');
-
-        await testSubjects.click('confirmModalConfirmButton');
-
-        await testSubjects.missingOrFail('custom-fields-list');
+      it('exposes the templates table', async () => {
+        await testSubjects.existOrFail('templates-table');
       });
     });
 
-    describe('Templates', function () {
-      before(async function () {
-        // The redesign settings page does not manage templates (moved to the templates v2 UI).
-        if (await cases.common.isRedesignEnabled()) {
-          this.skip();
-        }
+    describe('Field library page (v2)', function () {
+      before(async () => {
+        await navigateToConfigure();
+        const configureUrl = (await browser.getCurrentUrl()).split(/[?#]/)[0].replace(/\/$/, '');
+        await browser.get(`${configureUrl}/field_library`);
+        await header.waitUntilLoadingHasFinished();
       });
 
-      it('adds a template', async () => {
-        await testSubjects.existOrFail('templates-form-group');
-        await common.clickAndValidate('add-template', 'common-flyout');
-
-        await testSubjects.setValue('template-name-input', 'Template name');
-        await comboBox.setCustom('template-tags', 'tag-t1');
-        await testSubjects.setValue('template-description-input', 'Template description');
-
-        const caseTitle = await find.byCssSelector(
-          `[data-test-subj="input"][aria-describedby="caseTitle"]`
-        );
-        await caseTitle.focus();
-        await caseTitle.type('case with template');
-
-        await cases.create.setDescription('test description');
-
-        await cases.create.setTags('tagme');
-        await cases.create.setCategory('new');
-
-        await testSubjects.click('common-flyout-save');
-        expect(await testSubjects.exists('euiFlyoutCloseButton')).to.be(false);
-
-        await retry.waitFor('templates-list', async () => {
-          return await testSubjects.exists('templates-list');
-        });
-
-        expect(await testSubjects.getVisibleText('templates-list')).to.be('Template name\ntag-t1');
-      });
-
-      it('updates a template', async () => {
-        await testSubjects.existOrFail('templates-form-group');
-        const editButton = await find.byCssSelector('[data-test-subj*="-template-edit"]');
-
-        await editButton.click();
-
-        await testSubjects.setValue('template-name-input', 'Updated template name!');
-        await comboBox.setCustom('template-tags', 'tag-t1');
-        await testSubjects.setValue('template-description-input', 'Template description updated');
-
-        const caseTitle = await find.byCssSelector(
-          `[data-test-subj="input"][aria-describedby="caseTitle"]`
-        );
-        await caseTitle.focus();
-        await caseTitle.type('!!');
-
-        await cases.create.setDescription('test description!!');
-
-        await cases.create.setTags('case-tag');
-        await cases.create.setCategory('new!');
-
-        await testSubjects.click('common-flyout-save');
-        expect(await testSubjects.exists('euiFlyoutCloseButton')).to.be(false);
-
-        await retry.waitFor('templates-list', async () => {
-          return await testSubjects.exists('templates-list');
-        });
-
-        expect(await testSubjects.getVisibleText('templates-list')).to.be(
-          'Updated template name!\ntag-t1'
-        );
-      });
-
-      it('deletes a template', async () => {
-        await testSubjects.existOrFail('templates-form-group');
-        const deleteButton = await find.byCssSelector('[data-test-subj*="-template-delete"]');
-
-        await deleteButton.click();
-
-        await testSubjects.existOrFail('confirm-delete-modal');
-
-        await testSubjects.click('confirmModalConfirmButton');
-
-        await testSubjects.missingOrFail('template-list');
+      it('renders the field library page with its create affordance', async () => {
+        await testSubjects.existOrFail('fieldDefinitionsTable');
+        await testSubjects.existOrFail('createFieldDefinitionButton');
       });
     });
   });
