+++ x-pack/solutions/observability/test/serverless/functional/test_suites/screenshot_creation/response_ops_docs/cases/settings.ts
@@ -9,11 +9,13 @@ import { OBSERVABILITY_OWNER } from '@kbn/cases-plugin/common';
 import { navigateToCasesApp } from '@kbn/test-suites-xpack-platform/serverless/shared/lib/cases';
 import type { FtrProviderContext } from '../../../../ftr_provider_context';
 
-export default function ({ getPageObject, getPageObjects, getService }: FtrProviderContext) {
+export default function ({ getPageObject, getService }: FtrProviderContext) {
   const retry = getService('retry');
+  const browser = getService('browser');
   const svlCases = getService('svlCases');
   const svlCommonScreenshots = getService('svlCommonScreenshots');
   const svlCommonPage = getPageObject('svlCommonPage');
+  const header = getPageObject('header');
   const screenshotDirectories = ['response_ops_docs', 'observability_cases'];
   const testSubjects = getService('testSubjects');
   const cases = getService('cases');
@@ -28,6 +30,10 @@ export default function ({ getPageObject, getPageObjects, getService }: FtrProvi
     });
 
     it('case settings screenshots', async function () {
+      // With the templates feature flag pinned ON for this suite, custom fields and
+      // templates are managed on the dedicated v2 templates / field-library pages
+      // rather than inline on the Case Settings page. Capture the settings page,
+      // the templates list, the field library, and the add-connector flyout.
       await navigateToCasesApp(getPageObject, getService, owner);
       // The redesigned settings page drops the custom fields and templates management these
       // screenshots document, so skip while the redesign is on.
@@ -38,39 +44,46 @@ export default function ({ getPageObject, getPageObjects, getService }: FtrProvi
         return await testSubjects.exists('configure-case-button');
       });
       await testSubjects.click('configure-case-button');
-      await retry.waitFor('add-custom-field exist', async () => {
-        return await testSubjects.exists('add-custom-field');
+      await header.waitUntilLoadingHasFinished();
+      await retry.waitFor('case-configure-title exist', async () => {
+        return await testSubjects.exists('case-configure-title');
       });
-      await testSubjects.click('add-custom-field');
-      await svlCommonScreenshots.takeScreenshot(
-        'observability-cases-custom-fields',
-        screenshotDirectories,
-        1400,
-        700
-      );
-      await testSubjects.setValue('custom-field-label-input', 'my-field');
-      await retry.waitFor('common-flyout-save exist', async () => {
-        return await testSubjects.exists('common-flyout-save');
-      });
-      await testSubjects.click('common-flyout-save');
       await svlCommonScreenshots.takeScreenshot(
         'observability-cases-settings',
         screenshotDirectories
       );
-      await retry.waitFor('add-template exist', async () => {
-        return await testSubjects.isEnabled('add-template');
+
+      // Templates list page — reachable when the templates flag is ON.
+      // Strip any query string / hash so the sub-path is appended cleanly.
+      const configureUrl = (await browser.getCurrentUrl()).split(/[?#]/)[0].replace(/\/$/, '');
+      await browser.get(`${configureUrl}/templates`);
+      await header.waitUntilLoadingHasFinished();
+      await retry.waitFor('templates-table exist', async () => {
+        return await testSubjects.exists('templates-table');
       });
-      await testSubjects.click('add-template');
       await svlCommonScreenshots.takeScreenshot(
         'observability-cases-templates',
         screenshotDirectories,
         1400,
         1000
       );
-      await retry.waitFor('common-flyout-cancel exist', async () => {
-        return await testSubjects.exists('common-flyout-cancel');
+
+      // Field library page — reachable when the templates flag is ON.
+      await browser.get(`${configureUrl}/field_library`);
+      await header.waitUntilLoadingHasFinished();
+      await retry.waitFor('fieldDefinitionsTable exist', async () => {
+        return await testSubjects.exists('fieldDefinitionsTable');
       });
-      await testSubjects.click('common-flyout-cancel');
+      await svlCommonScreenshots.takeScreenshot(
+        'observability-cases-field-library',
+        screenshotDirectories,
+        1400,
+        700
+      );
+
+      // Add-connector flyout on the settings page (unchanged by the flag).
+      await browser.get(configureUrl);
+      await header.waitUntilLoadingHasFinished();
       await retry.waitFor('dropdown-connectors exist', async () => {
         return await testSubjects.exists('dropdown-connectors');
       });
